demo/adversary/requirements.txt
selenium
python-dotenv
numpy